+++ ponto.h
@@ -3,10 +3,10 @@
 
 typedef struct ponto Ponto;
 
-Ponto* criaPonto (char* nome, float X, float Y, int index);
+Ponto* criaPonto (char* nome, long double  X, long double Y, int index);
 char* retornaNome (Ponto* ponto);
-double retornaX (Ponto* ponto);
-double retornaY (Ponto* ponto);
+long double retornaX (Ponto* ponto);
+long double retornaY (Ponto* ponto);
 int retornaIndex(Ponto* ponto);
 void destroiPonto (Ponto* ponto);
